Комп практикум/ЛР10/requirements.txt
pillow>=12.2.0
huggingface_hub>=1.12.2
python-dotenv>=1.2.2
Flask>=3.1.3